Consolidate settings actions and improve layout
